simple-definition-list-blocks/details の構成オブジェクトを修正
